Perturbation-Metric-main/requirements.txt
captum==0.5.0
dataset==1.6.2
diffusers==0.18.2
einops==0.6.1
grad_cam==1.4.6
h5py==3.1.0
imageio==2.22.1
matplotlib==3.5.3
numpy==1.21.6
Pillow==9.2.0
scikit_learn==1.0.2
scipy==1.7.3
timm==0.4.12
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.30.2
accelerate==0.20.3